third_party/blink/web_tests/fast/loader/load-url-externally-user-gesture.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<a href='about:blank'></a>
<script>
    description('Check that a javascript generated mouse click is not enough to open a new tab.');

    document.querySelector('a').dispatchEvent(new MouseEvent('click', {ctrlKey: true, metaKey: true}));

    if (window.testRunner)
        shouldBe("testRunner.windowCount()", "1");
</script>
